core java for the impatient(chapter-3 exercises-13)
Q.Write a run method that takes an array of Runnable instances and
returns a Runnable whose run method executes them in order.Return a
lambda expression.

core java for the impatient(chapter-8, Q-13)

Q.Join all elements in a Stream<ArrayList<T>> to one ArrayList<T>.Show
how to do this with each of the 3 forms of reduce.